Created empty methods.
