+++ src/test/java/HammurabiTest.java
@@ -1,5 +1,3 @@
-package hammurabi;
-
 import static org.junit.Assert.*;
 
 import org.junit.Before;
